migrate to junit 5, extract benchmarks and memory test to e2e module
closes #107
