+++ fastexcel-reader/src/test/java/org/dhatim/fastexcel/reader/FastExcelReaderTest.java
@@ -24,7 +24,6 @@ import java.math.BigDecimal;
 import java.text.SimpleDateFormat;
 import java.util.*;
 import java.util.logging.Logger;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
@@ -35,7 +34,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class FastExcelReaderTest {
+class FastExcelReaderTest {
 
     private static final class RowDates {
 
@@ -43,7 +42,7 @@ public class FastExcelReaderTest {
         private final String date1;
         private final String date2;
 
-        public RowDates(int rowNum, String date1, String date2) {
+        RowDates(int rowNum, String date1, String date2) {
             this.rowNum = rowNum;
             this.date1 = date1;
             this.date2 = date2;
@@ -75,7 +74,7 @@ public class FastExcelReaderTest {
     private static final SimpleDateFormat FORMAT = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm");
 
     @Test
-    public void testDates() throws IOException {
+    void testDates() throws IOException {
         assertThat(readUsingFastExcel()).isEqualTo(readUsingPOI());
     }
 
@@ -139,7 +138,7 @@ public class FastExcelReaderTest {
             "/xlsx/write.xlsx",
             // "/xlsx/xlsx-stream-d-date-cell.xlsx"
     })
-    public void testFile(String file) {
+    void testFile(String file) {
         LOGGER.info("Test " + file);
         try (InputStream inputStream = open(file); InputStream inputStream2 = open(file)) {
             try (ReadableWorkbook excel = new ReadableWorkbook(inputStream); Workbook workbook = WorkbookFactory.create(inputStream2)) {
