+++ fastexcel-reader/src/test/java/org/dhatim/fastexcel/reader/HeaderTest.java
@@ -22,24 +22,24 @@ import java.io.InputStream;
 
 import org.junit.jupiter.api.Test;
 
-public class HeaderTest {
+class HeaderTest {
 
     @Test
-    public void testInsufficientBytes() {
+    void testInsufficientBytes() {
         byte[] bytes = new byte[1];
         assertThatThrownBy(() -> ReadableWorkbook.isOLE2Header(bytes)).isInstanceOf(IllegalArgumentException.class);
         assertThatThrownBy(() -> ReadableWorkbook.isOOXMLZipHeader(bytes)).isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
-    public void testOLE2File() throws IOException {
+    void testOLE2File() throws IOException {
         byte[] bytes = readFirstBytes("/xls/defined_names_simple.xls", 8);
         assertThat(ReadableWorkbook.isOLE2Header(bytes)).isTrue();
         assertThat(ReadableWorkbook.isOOXMLZipHeader(bytes)).isFalse();
     }
 
     @Test
-    public void testOOXMLZipFile() throws IOException {
+    void testOOXMLZipFile() throws IOException {
         byte[] bytes = readFirstBytes("/xlsx/defined_names_simple.xlsx", 8);
         assertThat(ReadableWorkbook.isOLE2Header(bytes)).isFalse();
         assertThat(ReadableWorkbook.isOOXMLZipHeader(bytes)).isTrue();
