+++ pom.xml
@@ -15,6 +15,7 @@
     <modules>
         <module>fastexcel-writer</module>
         <module>fastexcel-reader</module>
+        <module>e2e</module>
     </modules>
 
     <name>Fastexcel Project</name>
@@ -48,7 +49,22 @@
         <sonar.login>${env.SONAR_AUTH_TOKEN}</sonar.login>
         <sonar.github.repository>dhatim/fastexcel</sonar.github.repository>
         <sonar.github.oauth>${env.SONAR_GITHUB_OAUTH}</sonar.github.oauth>
+        <junit.jupiter.version>5.2.0</junit.jupiter.version>
+        <junit.platform.version>1.3.2</junit.platform.version>
     </properties>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>org.junit</groupId>
+                <artifactId>junit-bom</artifactId>
+                <version>${junit.jupiter.version}</version>
+                <scope>import</scope>
+                <type>pom</type>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
+
 
     <build>
         <plugins>
